requirements.txt
kagglehub==0.3.6
google-generativeai==0.7.2
pandas==2.2.2
langchain==0.2.6
langchain-google-genai==1.0.10
scikit-learn==1.5.0
python-dotenv==1.1.1
streamlit==1.36.0
